requirements.txt
accelerate==1.1.1
peft==0.17.0
transformers==4.46.3
huggingface_hub==0.34.4

diffusers @ git+https://github.com/huggingface/diffusers.git@1b48db4c8fe76ffffa7382fd74d9f04d54aa5a16

numpy==1.26.4
opencv_python==4.10.0.84
pillow==10.3.0
scipy==1.13.1
scikit-image==0.24.0
matplotlib==3.9.1
tqdm==4.66.4
cloudpickle==3.0.0
omegaconf==2.3.0
pycocotools==2.0.8
fvcore==0.1.5.post20221221
av==12.3.0
PyYAML==6.0.1

gradio==4.44.1
gradio_client==1.3.0
pydantic==2.10.4
fastapi==0.116.1
starlette>=0.40.0,<0.48.0

mediapipe==0.10.14
protobuf>=3.20.3,<5

# xformers>=0.0.27 ; sys_platform != "darwin" and platform_machine == "x86_64"
